Move BotController from main.ts to /process-controllers
